requirements.txt
python-telegram-bot==20.6
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
loguru==0.7.2